refactor: create a hook useCartList
