+++ .vscode/settings.json
@@ -3,6 +3,7 @@
   "cSpell.words": [
     "commercetools",
     "inputtext",
+    "Iuse",
     "primeflex",
     "primeicons",
     "primereact",
