+++ src/components/Links/Links.module.scss
@@ -1,4 +1,4 @@
-@import '../styles/global';
+@import '../../styles/global';
 
 .fontSize {
   font-size: 2em !important;
