+++ src/components/Links/LinksIsAuth.module.scss
@@ -1,0 +1,5 @@
+@import '../../styles/global';
+
+.fontSize {
+  font-size: 2em !important;
+}
