+++ src/components/LinksForHeader.tsx
@@ -1,6 +1,8 @@
 import { useContext } from 'react';
 import { AuthContext } from './authProvider';
-import { commonLinks, LinksIsNotAuth, LinksIsAuth } from './Links';
+import { commonLinks } from './Links/Links';
+import { LinksIsAuth } from './Links/LinksIsAuth';
+import { LinksIsNotAuth } from './Links/LinksIsNotAuth';
 
 export const LinksForHeader = (): JSX.Element => {
   const { isAuth } = useContext(AuthContext);
